// DotNetCorePayroll.App/src/app/shared/generated/model/companyAddressModel.ts
/**
 * Zenit Payroll API V1
 * Zenit Payroll Web API written in ASP.NET Core Web API
 *
 * OpenAPI spec version: v1
 * Contact: molefeq@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AddressModel } from './addressModel';


export interface CompanyAddressModel { 
    companyId?: number;
    physicalAddress?: AddressModel;
    postalAddress?: AddressModel;
}